+++ editor/src/main/java/org/nmox/studio/editor/typing/JsTypedBreakInterceptor.java
@@ -127,6 +127,7 @@ public class JsTypedBreakInterceptor implements TypedBreakInterceptor {
         @MimeRegistration(mimeType = "text/x-crystal", service = TypedBreakInterceptor.Factory.class),
         @MimeRegistration(mimeType = "text/css", service = TypedBreakInterceptor.Factory.class),
         @MimeRegistration(mimeType = "text/x-scss", service = TypedBreakInterceptor.Factory.class),
+        @MimeRegistration(mimeType = "text/x-sass", service = TypedBreakInterceptor.Factory.class),
         @MimeRegistration(mimeType = "text/x-less", service = TypedBreakInterceptor.Factory.class)
     })
     public static class Factory implements TypedBreakInterceptor.Factory {